bocr/requirements.txt
wandb>=0.12.2
numpy>=1.21.1
uvicorn>=0.11.7
opencv-python
fastapi>=0.65.2
python-multipart>=0.0.5
email-validator>=1.0.5
requests>=2.23.0
celery>=4.4.2
passlib[bcrypt]>=1.7.2
tenacity>=6.1.0
pydantic>=1.6.2
emails>=0.5.15
raven>=6.10.0
gunicorn>=20.0.4
jinja2>=2.11.3
psycopg2-binary>=2.8.5
alembic>=1.4.2
sqlalchemy>=1.3.16
pytest>=5.4.1
python-jose[cryptography]>=3.1.0
mypy>=0.770
black>=19.10b0
isort>=4.3.21
autoflake>=1.3.1
flake8>=3.7.9
pytest>=5.4.1
sqlalchemy-stubs>=0.3
pytest-cov>=2.8.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0+cpu
pandas==1.3.2
albumentations==1.0.3
numpy==1.21.1
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=8.0.0
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.41.0
# logging -------------------------------------
tensorboard>=2.4.1
# plotting ------------------------------------
seaborn>=0.11.0
thop  # FLOPs computation
